server: Implement event source put/get/rm
Includes a major refactor of existing code to share common bits between
components and event sources.  The db package was removed and the
relevant code moved into v1 (the two were tightly coupled).

maelctl includes new commands:

    maelctl es put
    maelctl es rm

The IDL was reworked so that the put calls accept instances of the
structs they're storing.